fix: use append for cookie headers and update cookie generation to use u64 for max-age calculation

+++ src/service/auth_service.rs
@@ -136,8 +136,8 @@ impl AuthService {
         Err(ServiceError::internal())
     }
 
-    pub fn generate_cookie(cookie_name: &str, token: String, day: Option<u8>) -> String {
-        let max_age = day.unwrap_or(1) * 24 * 60 * 60; // 6 days in seconds (259200)
+    pub fn generate_cookie(cookie_name: &str, token: String, day: Option<u64>) -> String {
+        let max_age = day.unwrap_or(1) * 24 * 60 * 60;
         let is_production = AppEnv::new() == AppEnv::Production;
         let secure_flag = if is_production { "; Secure" } else { "" };
